bot/requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
pydantic==2.9.1
pydantic-settings
PyYAML==6.0.3
